And James said "Let there be a protocol to ...
separate the data from the presentation". And the protocol he called
JSON. And that which was below the protocol he called Yesod and that
which was above the protocol he called Angular. And he saw that it was
good.

And there was morning, and there was night, the fourth day.